+++ CHANGELOG.md
@@ -1,27 +1,39 @@
 # Changelog
 All Notable changes to `flipboxdigital/hubspot` will be documented in this file
 
-## 1.0.0-beta.5 - 2017-7-14
+## Unreleased
+### Changed
+- Reverted log levels
+
+### Added
+- Spark log trait for separate log file logging
+- Debug Mode setting to lower log levels in a production environment.
+
+## 1.0.0-beta.6 - 2017-07-14
+### Changed
+- Log level for contact lists
+
+## 1.0.0-beta.5 - 2017-07-14
 ### Added
 - Adding response body logging for add and remove from contact lists.
 
-## 1.0.0-beta.4 - 2017-7-7
+## 1.0.0-beta.4 - 2017-07-07
 ### Added
 - Updated dependencies
 
-## 1.0.0-beta.3 - 2017-7-6
+## 1.0.0-beta.3 - 2017-07-06
 ### Added
 - Contact batch
 - Contact List add/remove
 
-## 1.0.0-beta.2 - 2017-6-23
+## 1.0.0-beta.2 - 2017-06-23
 ### Fixed
 - Type case issues.
 
-## 1.0.0-beta.1 - 2017-6-14
+## 1.0.0-beta.1 - 2017-06-14
 ### Added
 - Transformer enhancements
 
-## 1.0.0-beta - 2017-6-13
+## 1.0.0-beta - 2017-06-13
 ### Added
 - Initial release!